refactor: Extract common members to Super class
