--- src/main/java/rentcompany/AbstractCar.java
@@ -1,13 +1,0 @@
-package rentcompany;
-
-public class AbstractCar implements Car {
-    @Override
-    public int calculateFuel() {
-        return 0;
-    }
-
-    @Override
-    public String generateReport() {
-        return null;
-    }
-}
