+++ src/main/java/rentcompany/Vehicle.java
@@ -1,0 +1,6 @@
+package rentcompany;
+
+public interface Vehicle {
+    int calculateFuel();
+    String generateReport();
+}
